chore: split app-url utils module [YTFRONT-5734]
