+++ packages/ui/src/ui/utils/app-url/makeBundleUrl.ts
@@ -1,0 +1,11 @@
+import {Page} from '../../../shared/constants/settings';
+import {YT} from '../../config/yt-config';
+
+type Params = {
+    bundle: string;
+    cluster?: string;
+};
+
+export const makeBundleUrl = ({bundle, cluster}: Params): string => {
+    return `/${cluster || YT.cluster}/${Page.TABLET_CELL_BUNDLES}/instances?activeBundle=${bundle}`;
+};
